+++ packages/gatsby-cli/src/structured-errors/error-schema.ts
@@ -6,8 +6,8 @@ export const Position: Joi.ObjectSchema<ILocationPosition> = Joi.object().keys({
   column: Joi.number(),
 })
 
-export const errorSchema: Joi.ObjectSchema<IStructuredError> = Joi.object().keys(
-  {
+export const errorSchema: Joi.ObjectSchema<IStructuredError> =
+  Joi.object().keys({
     code: Joi.string(),
     text: Joi.string(),
     stack: Joi.array()
@@ -37,5 +37,4 @@ export const errorSchema: Joi.ObjectSchema<IStructuredError> = Joi.object().keys
     group: Joi.string(),
     panicOnBuild: Joi.boolean(),
     pluginName: Joi.string(),
-  }
-)
+  })