+++ packages/gatsby-plugin-gatsby-cloud/src/components/buttons/index.js
@@ -1,0 +1,12 @@
+import IndicatorButton from "./IndicatorButton"
+
+import LinkIndicatorButton from "./LinkIndicatorButton"
+import InfoIndicatorButton from "./InfoIndicatorButton"
+import GatsbyIndicatorButton from "./GatsbyIndicatorButton"
+
+export {
+  IndicatorButton,
+  LinkIndicatorButton,
+  InfoIndicatorButton,
+  GatsbyIndicatorButton,
+}