+++ packages/gatsby-plugin-offline/src/__tests__/gatsby-node.js
@@ -187,7 +187,8 @@ describe(`pluginOptionsSchema`, () => {
             handler: `StaleWhileRevalidate`,
           },
           {
-            urlPattern: /^https?:.*\.(png|jpg|jpeg|webp|svg|gif|tiff|js|woff|woff2|json|css)$/,
+            urlPattern:
+              /^https?:.*\.(png|jpg|jpeg|webp|svg|gif|tiff|js|woff|woff2|json|css)$/,
             handler: `StaleWhileRevalidate`,
           },
           {