+++ packages/gatsby/src/schema/infer/build-example-data.ts
@@ -205,8 +205,8 @@ const hasOnlyEmptyStrings = (descriptor: IValueDescriptor): boolean => {
 }
 
 const possibleTypes = (descriptor: IValueDescriptor = {}): Array<ValueType> =>
-  Object.keys(descriptor).filter(type => descriptor[type].total > 0) as Array<
-    ValueType
-  >
+  Object.keys(descriptor).filter(
+    type => descriptor[type].total > 0
+  ) as Array<ValueType>
 
 export { getExampleObject }