+++ packages/gatsby-admin/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.22.0-next.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.22.0-next.2...gatsby-admin@0.22.0-next.3) (2021-08-09)
+
+**Note:** Version bump only for package gatsby-admin
+
+# [0.22.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.22.0-next.1...gatsby-admin@0.22.0-next.2) (2021-08-05)
+
+**Note:** Version bump only for package gatsby-admin
+
+# [0.22.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.22.0-next.0...gatsby-admin@0.22.0-next.1) (2021-08-05)
+
+**Note:** Version bump only for package gatsby-admin
+
 # [0.22.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.21.0-next.6...gatsby-admin@0.22.0-next.0) (2021-07-29)
 
 **Note:** Version bump only for package gatsby-admin